Fixed memory leak in Aether based resolver.
When a session was created with (null), the code looked up the Deque for
the local repository.

However, when it was released, it was added to the Deque for the local repository
of the session. It was assumed that this would be the same, but it was a
LocalRepository with no type set in the create case and in the second case a
different LocalRepository was created by SimpleLocalRepositoryManager with the
type "simple". This caused the releaseSession method to put it to a different key,
than the createSession would use to look up.

This fix passes in "simple" as type, causing the two keys to be equal() and thus
the ConcurrentMap lookup to be successful.